https://issues.redhat.com/browse/WINDUP-3639
