+++ src/lib/IDE/CodeReadyStudio.py
@@ -1,0 +1,9 @@
+from src.lib.application import Application
+
+
+class CodeReadyStudio(Application):
+    """
+    Class for managing RH Code Ready Studio application
+    """
+
+    pass
