+++ src/lib/IDE/Eclipse.py
@@ -1,0 +1,9 @@
+from src.lib.application import Application
+
+
+class Eclipse(Application):
+    """
+    Class for managing Eclipse application
+    """
+
+    pass
