+++ src/lib/application.py
@@ -376,7 +376,7 @@ class Application(Desktop):
         """
         if target == "eap7":
             story_point_locator = "eap7_story_points.png"
-        elif target == "quarkus1":
+        elif target == "quarkus":
             story_point_locator = "quarkus_story_points.png"
         elif target == "eapxp":
             story_point_locator = "eapxp_story_points.png"
@@ -401,305 +401,3 @@ class Application(Desktop):
         """
         for i in range(0, position):
             self.press_keys("down")
-
-
-class CodeReadyStudio(Application):
-    """
-    Class for managing RH Code Ready Studio application
-    """
-
-    pass
-
-
-class Eclipse(Application):
-    """
-    Class for managing Eclipse application
-    """
-
-    pass
-
-
-class VisualStudioCode(Application):
-    """
-    Class for managing VSCode application
-    """
-
-    def close_ide(self):
-        """
-        Closes the IDE
-        """
-        self.press_keys("ctrl", "q")
-
-    def is_open_mta_perspective(self):
-        """
-        Checks if MTA perspective is already opened in VS Code
-
-        Returns:
-            (bool): True or False
-        """
-        try:
-            self.wait_find_element(locator_type="image", locator="mta_config_active.png")
-            return True
-        except Exception as exc:
-            logging.debug(
-                "An error occured while finding \
-                MTA perspective tab ! {}".format(
-                    str(exc),
-                ),
-            )
-            if "No matches found" in str(exc):
-                return False
-            else:
-                raise Exception(exc)
-
-    def open_mta_perspective(self):
-        """
-        Opens MTA perspective in VSCode IDE
-        """
-        if self.is_open_mta_perspective():
-            logging.info("MTA perspective is already opened !")
-            return
-        else:
-            # Click on the MTA icon in left sidebar
-            self.click_element(locator_type="image", locator="mta_config_inactive.png")
-
-    def run_simple_analysis(self, project, migration_target, packages=[]):
-        """
-        Runs analysis by adding the project and/or packages passed as argument
-
-        Args:
-            project (str): Full name of project to be analysed
-            packages (list): List of packages to be added to analysis
-
-        Returns:
-            None
-
-        Steps:
-            1) Click on MTA Configuration icon
-            2) Click new config(+) icon
-            3) Type project name in source
-            4) Click on config name and run analysis
-            5) Confirm analysis has started
-        """
-        time.sleep(20)
-        # Create new analysis configuration by clicking 'New Configuration' button
-        self.wait_find_element(locator_type="image", locator="create_new_config.png")
-        self.click_element(locator_type="image", locator="create_new_config.png")
-        # Wait for configuration page to load
-        time.sleep(3)
-
-        # Add project info to configuration
-        add_project_locator = self.image_locator("add_button.png")
-        add_project_buttons = self.find_elements(add_project_locator)
-        # Click the first match out of the two same buttons found
-        self.click(add_project_buttons[0])
-        self.type_text(text=project, enter=True)
-        self.press_keys("page_down")
-
-        # Add migration target info to configuration
-        # Default target is eap7. Uncheck 'eap7' if target is something else.
-        if migration_target != "eap7":
-            self.click_element(locator_type="image", locator="eap7_target_checked.png")
-            self.click_element(locator_type="image", locator="add_button.png")
-            self.type_text(text=migration_target, enter=True)
-
-        # Run analysis after clicking on config name
-        try:
-            self.click_element(locator_type="image", locator="config_name_region.png")
-        except Exception as exc:
-            if re.match(r"Found [0-9] matches+", str(exc)):
-                config_region = self.image_locator("config_name_region.png")
-                config_region_circles = self.find_elements(config_region)
-                self.click(config_region_circles[-1])
-        self.click_element(locator_type="image", locator="run_analysis.png")
-
-        # Verify analysis has started
-        self.wait_find_element(locator_type="image", locator="analysis_progress.png", timeout=240.0)
-
-    def is_analysis_complete(self):
-        """
-        Checks if run analysis has been completed
-
-        Returns:
-            (bool): True if analysis was completed
-        """
-        try:
-            self.wait_find_element(
-                locator_type="image",
-                locator="analysis_complete.png",
-                timeout=120.0,
-            )
-            return True
-        except Exception:
-            return False
-
-    def open_report_page(self):
-        self.click_element(locator_type="image", locator="open_report_button.png")
-        self.wait_find_element(locator_type="image", locator="report_page_header.png")
-
-
-class Intellij(Application):
-    """
-    Class for managing IntelliJ application
-    """
-
-    def close_ide(self):
-        """
-        Closes the IDE
-        """
-        self.click_element(locator_type="image", locator="file_menu.png")
-        self.press_keys("up")
-        self.press_keys("enter")
-        time.sleep(1)
-        self.press_keys("enter")
-        # try:
-        #     self.wait_find_element(locator_type="image", locator="confirm_exit.png", timeout=5.0)
-        #     self.click_element(locator_type="image", locator="exit_button.png")
-        # except Exception:
-        #     logging.info("No exit confirmation dialog found !")
-
-    def is_open_mta_perspective(self):
-        """
-        Checks if MTA perspective is already opened in IntelliJ
-
-        Returns:
-            (bool): True or False
-        """
-        try:
-            self.wait_find_element(locator_type="image", locator="mta_perspective_active.png")
-            return True
-        except Exception as exc:
-            try:
-                self.wait_find_element(
-                    locator_type="image",
-                    locator="mta_perspective_active_alt.png",
-                )
-                return True
-            except Exception:
-                logging.debug(
-                    "An error occured while finding \
-                    MTA perspective tab ! {}".format(
-                        str(exc),
-                    ),
-                )
-            if "No matches found" in str(exc):
-                return False
-            else:
-                raise Exception(exc)
-
-    def open_mta_perspective(self):
-        """
-        Opens MTA perspective in IntelliJ
-        """
-        if self.is_open_mta_perspective():
-            logging.info("MTA perspective is already opened !")
-            return
-        else:
-            # Click on the MTA tab in left sidebar
-            self.click_element(locator_type="image", locator="mta_tab.png")
-
-    def run_simple_analysis(self, project, migration_target, packages=[]):
-        # todo: set the migration_target to be a list instead of string
-        """
-        Runs analysis by adding the project and/or packages passed as argument
-
-        Args:
-            project (str): Full name of project to be analysed
-            migration_target (str): Target technology for migration
-            packages (list): List of packages to be added to analysis
-
-        Returns:
-            None
-
-        Steps:
-            1) Click on MTA Configuration tab
-            2) Right click anywhere and create new configuration
-            3) Type project name in source
-            4) Provide mta cli path, if not present
-            5) Select the target technology
-            6) Right click on config name and run
-            7) Confirm analysis has started
-        """
-        try:
-            self.click_element(locator_type="image", locator="console_opened.png")
-            self.press_keys("shift", "esc")
-        except Exception as exc:
-            try:
-                self.click_element(locator_type="image", locator="console_opened_alt.png")
-                self.press_keys("shift", "esc")
-            except Exception:
-                logging.debug("Console is already closed !")
-            logging.debug("Console is not opened ! {}".format(str(exc)))
-        config_create_region = self.two_coordinate_locator(
-            locator_type="point",
-            x_coordinate=110,
-            y_coordinate=370,
-        )
-        self.click(config_create_region)
-        self.click(action="right_click")
-        self.press_keys("down")
-        self.press_keys("enter")
-        time.sleep(3)
-        self.wait_find_element(locator_type="image", locator="mta_config_page_opened.png")
-        # Region defining the configuration name
-        config_name_region = self.define_region(778, 248, 1125, 284)
-        add_project_locator = self.image_locator("add_project_button.png")
-        add_project_buttons = self.find_elements(add_project_locator)
-        # Provide mta cli path
-        try:
-            # self.click_element(locator_type="image", locator="mta_cli_input.png")
-            self.click(add_project_buttons[0])
-            time.sleep(2)
-            for letter in config_data["mta_cli_path"]:
-                self.type_text(letter)
-            self.press_keys("enter")
-        except Exception as exc:
-            logging.debug("MTA cli path is already present ! {}".format(exc))
-        if migration_target != "eap7":
-            try:
-                self.wait_find_element(locator_type="image", locator="eap7_checked.png")
-                self.click_element(locator_type="image", locator="eap7_checked.png")
-            except Exception as exc:
-                logging.debug("Default target eap7 is not checked ! {}".format(str(exc)))
-
-        # Click the first match out of the two same buttons found
-        self.click(add_project_buttons[1])
-        self.type_text(text=project, enter=True)
-        self.click(config_name_region)
-        print(f'buttons found: {str(len(add_project_buttons))}')
-        # todo: find why it only finds 2 elements not 3
-        if len(add_project_buttons) != 3:
-            self.press_keys("page_down")
-        self.click_element(locator_type="image", locator="add_project_button.png")
-        self.type_text(text=migration_target, enter=True)
-        config_run_region = self.two_coordinate_locator(
-            locator_type="point",
-            x_coordinate=110,
-            y_coordinate=870,
-        )
-        self.click(config_run_region)
-        # Search for configuration name that has to be run
-        self.type_text("configuration1")
-        self.press_keys("enter")
-        # Find config name highlighted and select correct config if multiple matches are found
-        self.wait_find_element(locator_type="image", locator="config_name_highlighted.png")
-        self.click(action="right_click")
-        self.press_keys("up")
-        self.press_keys("enter")
-        # Wait for analysis to be completed in IDE terminal
-        self.wait_find_element(
-            locator_type="image",
-            locator="analysis_complete_terminal.png",
-            timeout=120.0,
-        )
-        # Select the run configuration and open report page in browser
-        try:
-            self.click_element(locator_type="image", locator="mta_perspective_active.png")
-        except Exception:
-            self.click_element(locator_type="image", locator="mta_perspective_active_alt.png")
-        self.type_text("configuration1")
-        self.press_keys("up")
-        self.click_element(locator_type="image", locator="open_details_toggle.png")
-        self.click_element(locator_type="image", locator="report_selector.png")
-        # Verify the report page is opened
-        self.wait_find_element(locator_type="image", locator="report_page_header.png")
