+++ src/tests/test_analysis_intellij.py
@@ -14,7 +14,8 @@ def test_analysis_eap(setup_intellij):
     time.sleep(3)
     # todo: set the projects selection to be done dynamically instead of manually adding to config
     intellij.run_simple_analysis(
-        project=config_data["project_paths"]["eap7_generic"], migration_target="eap7",
+        project=config_data["project_paths"]["eap7_generic"],
+        migration_target="eap7",
     )
     assert intellij.is_analysis_complete()
     assert intellij.verify_story_points(target="eap7")
@@ -30,7 +31,8 @@ def test_analysis_eapxp(setup_intellij):
     # Intellij freezes without this sleep
     time.sleep(3)
     intellij.run_simple_analysis(
-        project=config_data["project_paths"]["eapxp_ruleset"], migration_target="eapxp",
+        project=config_data["project_paths"]["eapxp_ruleset"],
+        migration_target="eapxp",
     )
     assert intellij.is_analysis_complete()
     assert intellij.verify_story_points(target="eapxp")
@@ -46,7 +48,8 @@ def test_analysis_quarkus(setup_intellij):
     # Intellij freezes without this sleep
     time.sleep(3)
     intellij.run_simple_analysis(
-        project=config_data["project_paths"]["quarkus_ruleset"], migration_target="quarkus",
+        project=config_data["project_paths"]["quarkus_ruleset"],
+        migration_target="quarkus",
     )
     assert intellij.is_analysis_complete()
     assert intellij.verify_story_points(target="quarkus")
